requirements.txt
click>=8.0
pyyaml>=6.0
pytest>=7.0
